Update UidInspectorProperty.cs
Added In Drag and drop for files. Simply drag a file like a scene onto any part of the UID property and it will add the UID for that file. Will use the first file when multiple files are dragged.

This plugin was just what I needed having the same problem with wanting to use UIDs. I added Drag and drop and thought best to see if its wanted to be merged into the original.